requirements.txt
django-floppyforms==1.8.0
django-category==2.0.1
djangorestframework==3.10.3
djangorestframework-gis==0.14